factor out configdrive handling code from secstor
