+++ config-drive/src/main/java/org/apache/cloudstack/storage/StorageAttacher.java
@@ -1,0 +1,27 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+package org.apache.cloudstack.storage;
+
+import java.net.URI;
+
+public interface StorageAttacher {
+    String getRootDir(String secUrl, Integer nfsVersion);
+
+    void umount(String localRootPath, URI uri);
+
+    void mount(String localRootPath, String remoteDevice, URI uri, Integer nfsVersion);
+}
